add:投稿する仕組みを実装（スタンプ選択以外） #10
