+++ src/components/room/RoomNavbar.js
@@ -2,11 +2,12 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import CancelDialog from "./CancelDialog";
 import SelectStamp from "./SelectStamp";
-import { setSelectionRange } from "@testing-library/user-event/dist/utils";
+import SetBathGoal from "./SetBathGoal";
 
 const RoomNavbar = ({ lastPostType }) => {
   const [showSelectStamp, setShowSelectStamp] = useState(false);
   const [showCancelDialog, setShowCancelDialog] = useState(false);
+  const [showSetBathGoal, setShowSetBathGoal] = useState(false);
   const [showPointUp, setShowPointUp] = useState(false);
   const [showLevelUp, setShowLevelUp] = useState(false);
 
@@ -28,6 +29,14 @@ const RoomNavbar = ({ lastPostType }) => {
 
   const closeCancelDialog = () => {
     setShowCancelDialog(false);
+  };
+
+  const openSetBathGoal= () => {
+    setShowSetBathGoal(true);
+  };
+
+  const closeSetBathGoal = () => {
+    setShowSetBathGoal(false);
   };
 
   return (
@@ -56,7 +65,8 @@ const RoomNavbar = ({ lastPostType }) => {
         )}
       </div>
       {showCancelDialog && <CancelDialog closeCancelDialog={closeCancelDialog}/>}
-      {showSelectStamp && <SelectStamp closeSelectStamp={closeSelectStamp} lastPostType={lastPostType}/>}
+      {showSelectStamp && <SelectStamp closeSelectStamp={closeSelectStamp} lastPostType={lastPostType} openSetBathGoal={openSetBathGoal} />}
+      {showSetBathGoal && <SetBathGoal closeSetBathGoal={closeSetBathGoal} closeSelectStamp={closeSelectStamp}/>}
     </div>
   );
 };
